# mudeopensrvapi/src/yaml/mudeopenapiws.yaml
...
swagger: '2.0'
info:
  title: MUDEOPEN WS-API 10-01-2023
  description: API per i WS REST
  version: 1.0.0
host: 'localhost:8080'
schemes:
  - http
basePath: /mudeopen/ws/rest
definitions:
...
  GeoCellula:
    properties:
      chiaveCarotaggio:
          type: string
      codCellula:
          format: int64
          type: integer
      dataGeoriferimento:
          format: date
          type: string
      GeoFabbricato:
        items:
          $ref: '#/definitions/GeoFabbricato'
        type: array
    type: object
  Allegato:
    properties:
      idTipoallegato:
          format: int64
          type: integer
      idAllegatoFile:
          format: int64
          type: integer
      nomeAllegato:
          type: string
    type: object
  GeoUbicazione:
    properties:
      bis:
          type: string
      bisinterno:
          type: string
      cap:
          type: string
      descFonteUbicazione:
          type: string
      descVia:
          type: string
      flPersonalizzato:
          type: string
      flPrincipale:
          type: string
      idCivicoTopon:
          format: int64
          type: integer
      idViaTopon:
          format: int64
          type: integer
      interno:
          type: string
      interno2:
          type: string
      numCivico:
          type: string
      piano:
          type: string
      scala:
          type: string
      secondario:
          type: string
      sedime:
          type: string
    type: object
  CSIGeometry:
    properties:
      geometria:
          type: string
    type: object
  MudeRVersioneModello:
    properties:
      codModello:
          type: string
      codVersioneModello:
          type: string
    type: object
  IstanzaMude:
    properties:
      xmlStream:
          type: string
      GeoRiferimento:
        items:
          $ref: '#/definitions/GeoRiferimento'
        type: array
      GeoUbicazione:
        items:
          $ref: '#/definitions/GeoUbicazione'
        type: array
      Allegato:
        items:
          $ref: '#/definitions/Allegato'
        type: array
      mudeRVersioneModello:
        $ref: '#/definitions/MudeRVersioneModello'
    type: object
  GeoRiferimento:
    properties:
      idLivelloPoligono:
          format: int64
          type: integer
      descLivelloPoligono:
          type: string
      servizioFonte:
          format: int64
          type: integer
      servizioLivello:
          format: int64
          type: integer
      dataGeoriferimento:
          format: date
          type: string
      codIstatComune:
          type: string
      GeoCellula:
        items:
          $ref: '#/definitions/GeoCellula'
        type: array
      GeoCatasto:
        items:
          $ref: '#/definitions/GeoCatasto'
        type: array
      GeoDatocarota:
        items:
          $ref: '#/definitions/GeoDatocarota'
        type: array
      GeoUbicazione:
        items:
          $ref: '#/definitions/GeoUbicazione'
        type: array
      csiGeometry:
        $ref: '#/definitions/CSIGeometry'
    type: object
  GeoDatocarota:
    properties:
      tipoCarotaggio:
          type: string
      idFonte:
          format: int64
          type: integer
      descFonte:
          type: string
      idLivello:
          format: int64
          type: integer
      dataGeoriferimento:
          format: date
          type: string
      descLivello:
          type: string
      valore:
          type: string
    type: object
  GeoCatasto:
    properties:
      foglio:
          type: string
      particella:
          type: string
      subalterno:
          type: string
      flTipoCatasto:
          type: string
      flPersonalizzato:
          type: string
      chiaveCarotaggio:
          type: string
      descFonteCatasto:
          type: string
    type: object
  GeoFabbricato:
    properties:
      chiaveCarotaggio:
          type: string
      codFabbricato:
          format: int64
          type: integer
      GeoUbicazione:
        items:
          $ref: '#/definitions/GeoUbicazione'
        type: array
      GeoCatasto:
        items:
          $ref: '#/definitions/GeoCatasto'
        type: array
    type: object
##################################
####
##################################    
...
  AllegatoIstanzaVO:
    properties:
      numeroIstanza:
        type: string
      tipoAllegato:
        type: string
      nomeFile:
        type: string
    type: object
  DatiSintesiIstanzaVO:
    type: object
    properties:
      numeroIstanza:
        type: string
      tipoIstanza:
        type: string
      speciePratica:
        type: string
      tipoIntestatario:
        type: string
      cognome:
        type: string
      nome:
        type: string
      ragioneSociale:
        type: string
      professionista:
        type: string
      sedime:
        type: string
      descrizione:
        type: string
      numeroCivico:
        type: string
      bis:
        type: string
      interno:
        type: string
      comune:
        type: string
      statoIstanza:
        type: string
      istanzaDiRiferimento:
        type: string
      numeroIntervento:
        type: string
...
  TipoTracciatoXML:
    properties:
      codTipoTracciato:
        type: string
      descrizione:
        type: string
      xmlPresente:
        type: boolean
    type: object
  IstanzaEstesaVO:
    properties:
      numeroMude:
        type: string
      tipoIstanza:
        type: string
      codiceTipoIstanza: 
        type: string
      speciePratica:
        type: string
      identificativoPratica: 
        type: string
      tipoIntestatario:
        type: string
      intestatario:
        type: string
      nome:
        type: string
      numeroCivico:
        type: string
      bis:
        type: string
      interno:
        type: string
      sedime:
        type: string
      via:
        type: string
      codiceSpeciePratica:
        type: string
      numeroIntervento:
        type: string
      statoIstanza: 
        type: string
      codiceStatoIstanza:
        type: string
      occupazioneSuoloPubblico:
        type: boolean
      dataRicezione:
        format: date
        type: string
      dataUltimaVariazioneStato:
        format: date
        type: string
      tipiTracciatiXML:
        $ref: '#/definitions/TipoTracciatoXML'
    type: object
...
  DefaultResponse:
    properties:
      code:
        type: string
      detail:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
    type: object
  ErrorResponse:
    properties:
      status:
        type: string
      code:
        type: string
      title:
        type: string
      detail:
        type: object
        additionalProperties:
          type: string
      links:
        type: array
        items:
          type: string
    type: object
#
# 2CHECK
#
...
  TraccXMLVO:
    properties:
      tracciatoXML:
        type: string
      versione:
        type: string
      versioneTracciato:
        type: string
    type: object
  IstanzaExtVO:
    properties:
      codIstatComune:
          type: string
      tipologiaIstanza:
          type: string
      numeroIstanza:
          type: string
      numeroFascicoloIntervento:
          type: string
      speciePratica:
          type: string
      annoIntervento:
          format: int64
          type: integer
      annoIstanza:
          format: int64
          type: integer
      tipoIntestatario:
          type: string
      intestatarioCognome:
          type: string
      intestatarioNome:
          type: string
      intestatarioRagioneSociale:
          type: string
      professionistaCognome:
          type: string
      professionistaNome:
          type: string
      ubicazioneSedime:
          type: string
      ubicazioneDescVia:
          type: string
      ubicazioneNumeroCivico:
          format: int64
          type: integer
      ubicazioneBis:
          type: string
      ubicazioneInterno:
          type: string
      operePrecarie:
          type: string
      numeroMudeIstanzaPrincipale:
          type: string
      compilanteCodiceFiscale:
          type: string
      tracciatiXML:
          type: array
          items:
            $ref: '#/definitions/TraccXMLVO'
      timestampInserimento:
          type: string
    type: object
...
  CambioStatoIstanzaVO:
    type: object
    properties:
      data_cambio_stato:
        type: string
        format: datetime
      oggetto_notifica:
        type: string
      messaggio_notifica:
        type: string
      numero_protocollo:
        type: string
      data_protocollo:
        type: string
        format: datetime
  Geometry:
    type: object
    properties:
      value:
        type: String
        enum: [LineString,Polygon,MultiPolygon,MultiPoint]
  EditedFeature:
    type: object
    properties:
      type:
        type: string
      id:
        format: int64
        type: integer
      geometry:
        $ref: '#/definitions/Geometry'
      properties:
        type: object
        additionalProperties:
          type: string
paths:
  /istanze:
    get:
      operationId: ricercaPaginataIstanze03
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - format: int32
          in: query
          name: page
          required: true
          type: integer
        - format: int32
          in: query
          name: size
          required: true
          type: integer
        - in: query
          name: filter
          type: string
      produces:
        - application/json
      responses:
        '200':
          headers:
            X-Page:
              type: integer
              format: int32
              description: For pagination purposes. The current page.
            X-Page-Size:
              type: integer
              format: int32
              description: For pagination purposes. The page size.
            X-Total-Elements:
              type: integer
              format: int32
              description: For pagination purposes. The total number of elements satisfying the search query.
            X-Total-Pages:
              type: integer
              format: int32
              description: For pagination purposes. The total number of pages for representing the total number of elements satisfying the query.
          description: successful operation
          schema:
            items:
              $ref: '#/definitions/IstanzaEstesaVO'
            type: array
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Index not found
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
...
  /istanza/{numeroIstanza}:
    get:
      operationId: ricercaDatiSintesiIstanza
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: numeroIstanza
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: successful response
          schema:
            $ref: '#/definitions/DatiSintesiIstanzaVO'
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Index not found
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
...
  /file-istanza/{numeroIstanza}:
    get:
      operationId: EstraiFileIstanza
      produces:
        - application/octet-stream
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: numeroIstanza
          required: true
          type: string
      responses:
        '200':
          description: successful response
          schema:
            type: file
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
  /allegato-istanza/{numeroIstanza}:
    get:
      operationId: EstraiAllegatoIstanza
      produces:
        - application/octet-stream
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: numeroIstanza
          required: true
          type: string
        - in: query
          name: nomeFileAllegato
          required: true
          type: string
        - in: query
          name: sbustato
          required: true
          type: string
      responses:
        '200':
          description: successful response
          schema:
            type: file
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
  /elenco-allegati/{numeroIstanza}:
    get:
      operationId: ricercaElencoAllegati
      produces:
        - application/json
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: numeroIstanza
          required: true
          type: string
      responses:
        '200':
          description: successful response
          schema:
            items:
              $ref: '#/definitions/AllegatoIstanzaVO'
            type: array
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Index not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
...
  /istanza:
    post:
      consumes:
        - application/json
      operationId: invioIstanza
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: body
          name: body
          schema:
            $ref: '#/definitions/IstanzaExtVO'
      produces:
        - application/json
      responses:
        '201':
          description: successful response
          schema:
            type: boolean
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
  /dati-protocollazione-istanza/{numeroIstanza}:
    get:
      operationId: visualizzaDatiProtocollazioneIstanza
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: numeroIstanza
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: successful response
          schema:
            $ref: '#/definitions/CambioStatoIstanzaVO'
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Not found
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
  /genera-numero-mude/{istatComune}/{nuovoIntervento}:
    post:
      consumes:
        - application/json
      operationId: generaNumeroMUDE
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: istatComune
          required: true
          type: string
        - in: path
          name: nuovoIntervento
          required: true
          type: string
      produces:
        - application/json
      responses:
        '201':
          description: successful response
          schema:
            type: boolean
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
########################
##
########################
  /get-elenco-allegati/{numeroIstanza}:
    get:
      consumes:
        - application/json
      operationId: getElencoAllegatiByNumeroIstanza
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: numeroIstanza
          type: string
          required: true
      produces:
        - application/json
      responses:
        '200':
          description: successful response
          schema:
            items:
              $ref: '#/definitions/Allegato'
            type: array
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
  /estrai-istanza/{numeroIstanza}:
    get:
      consumes:
        - application/json
      operationId: estraiIstanzaMude
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: numeroIstanza
          type: string
          required: true
      produces:
        - application/json
      responses:
        '200':
          description: successful response
          schema:
            $ref: '#/definitions/IstanzaMude'
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
  /elenco-geo-riferimento/{numeroIstanza}:
    get:
      consumes:
        - application/json
      operationId: getElencoGeoRiferimento
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: numeroIstanza
          type: string
          required: true
      produces:
        - application/json
      responses:
        '200':
          description: successful response
          schema:
            items:
              $ref: '#/definitions/GeoRiferimento'
            type: array
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error               
  /token/{codiceFruitore}:
    get:
      operationId: autenticazioneMUDE
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: codiceFruitore
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: successful response
          schema:
            type: string
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Not found
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
##################################
#### GEECO
##################################    
  /geeco/{environment}/{id_editing_session}/layers/{layer_id}:
    post:
      consumes:
        - application/json
      operationId: insertFeatureForEditingLayer
      parameters:
        - in: path
          name: environment
          required: true
          type: string
        - in: path
          name: layer_id
          required: true
          type: string
        - in: path
          name: id_editing_session
          required: true
          type: string
        - in: body
          name: body
          schema:
            $ref: '#/definitions/EditedFeature'
      produces:
        - application/json
      responses:
        '200':
          description: successful response
          schema:
            $ref: '#/definitions/EditedFeature'
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Index not found
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error